hackdeck/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Streamlit Frontend
streamlit==1.30.0

# Environment and Config
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1

# GitHub Integration
gitpython==3.1.41

# Codebase Analysis
codebase-digest==1.0.0

# LLM - Google Gemini
google-generativeai==0.3.2

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Utilities
python-slugify==8.0.1
validators==0.22.0